requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.2.2
django-bootstrap-v5==1.0.11
django-cors-headers==4.1.0
django-registration==3.3
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
openapi-codec==1.3.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python-dotenv==1.0.0
python-http-client==3.3.7
python-jwt==4.0.0
pytz==2023.3
requests==2.31.0
sendgrid==6.10.0
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
whitenoise==6.5.0
wrapt==1.15.0
